chess_game/requirements.txt
autogen>=0.2.0
python-chess>=1.0.0
PyYAML>=6.0
typing-extensions>=4.0.0